% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{checkFeatureIds}
\alias{checkFeatureIds}
\title{detect whether gene expression matrix is labeled with Entrez gene ids or Symbols}
\usage{
checkFeatureIds(sdat, minoverlapFrac = 0.25)
}
\arguments{
\item{sdat}{gene expression matrix that will be used for CMS assignment (default: none)}

\item{minoverlapFrac}{minimum overlap fraction required between rownames of expression matrix and either Entrez IDs or gene Symbols used for classification}
}
\value{
String specifying whether Entrez IDs or Symbols should be used for classification
}
\description{
detect whether gene expression matrix is labeled with Entrez gene ids or Symbols
}
\author{
Robert Piskol
}
