% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_classifier.R
\name{classifyCMString}
\alias{classifyCMString}
\title{classification of sample gene expression profiles into consensus molecular subtypes (CMS)}
\usage{
classifyCMString(x, applylog = TRUE, mediancenter = TRUE, zScore = TRUE,
  s.stat = "lambda.min", thresh = NULL)
}
\arguments{
\item{x}{ExpressionSet object or gene expression matrix for samples (rows) and genes (columns) [default: none]}

\item{applylog}{boolean specifying whether log2 transformation should be applied to the data [default: TRUE]}

\item{mediancenter}{boolean specifying whether expression of single genes should be median centered [default: TRUE]}

\item{zScore}{boolean specifying whether gene expression should be z-scored [default: TRUE]}

\item{s.stat}{Value of the penalty parameter lambda at which prediction will be made [default: lambda.min]}

\item{thresh}{minimum probability necessary for CMS call [default: NULL - the CMS with highest probability will be reported regardless of its probability]}
}
\value{
list with results
}
\description{
classification of sample gene expression profiles into consensus molecular subtypes (CMS)
}
\author{
Robert Piskol
}
