% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_data.R
\docType{data}
\name{exprs.CIT}
\alias{exprs.CIT}
\title{CIT data set}
\format{An ExpressionSet with 20192 features and 566 samples}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE39582}
}
\usage{
exprs.CIT
}
\description{
A dataset containing gene expression data for 566 samples from patients with early stage CRC.
The data is available from GEO ID GSE39582
}
\keyword{datasets}
