% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_data.R
\docType{data}
\name{exprs.crc.celllines}
\alias{exprs.crc.celllines}
\title{CRC cell line data set}
\format{An ExpressionSet with 30727 features and 522 samples}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=NNN}
}
\usage{
exprs.crc.celllines
}
\description{
A dataset containing gene expression data for 40 samples (4 cell lines, 2 implantation locations, 5 replicates)
The data is available from GEO IDs NNN
}
\keyword{datasets}
