% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_data.R
\docType{data}
\name{exprs.MVRM}
\alias{exprs.MVRM}
\title{MVRM data set}
\format{An ExpressionSet with 20192 features and 522 samples}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17536}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE17537}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE14333}
}
\usage{
exprs.MVRM
}
\description{
A dataset containing gene expression data for 522 samples from patients with early stage CRC.
The data is available from GEO IDs GSE17536, GSE17537, GSE14333
}
\keyword{datasets}
