% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_data.R
\docType{data}
\name{exprs.Procured}
\alias{exprs.Procured}
\title{Procured patient data set}
\format{An ExpressionSet with 819 features and 312 samples}
\usage{
exprs.Procured
}
\description{
A dataset containing gene expression data for 312 samples (182 primary, 130 met) that were profiled using NanoString
}
\keyword{datasets}
