% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_plotting.R
\name{figClassifyCMS}
\alias{figClassifyCMS}
\title{plot heatmap}
\usage{
figClassifyCMS(x, pred, clu.pred, sdat.sig, gclu.f, nam.ord,
  missing.genes = NULL, cex.axis.labs = 1, mutmat = NULL,
  gene.labs = TRUE, samp.labs = FALSE)
}
\arguments{
\item{x}{list with classification results (pred, clu.pred, sdat.sig, gclu.f, nam.ord, missing.genes) [default: none]}

\item{pred}{matrix of class prediction probabilities with samples in rows and classes in the columns.  [default: none]}

\item{clu.pred}{vector with predicted classes [default: none]}

\item{sdat.sig}{matrix with gene expression values for the classifier genes with samples in genes in rows and samples in columns [default: none]}

\item{gclu.f}{hierarchical clustering of classifier genes that specifies row ordering [default: none]}

\item{nam.ord}{vector of sample IDs that specify column ordering of heatmap [default: none]}

\item{missing.genes}{vector of genes that will be removed from the heatmap. These are usually genes that were not present in the expression data, but required by the classifier. However any other list can of IDs can be specified [default: NULL]}

\item{cex.axis.labs}{numeric scaling factor for axis labels (genes, samples) [default: 1]}

\item{mutmat}{matrix of numeric values that will be translated into colors and attached to the bottom of the heatmap. This can contain mutational information or any other phenotypic information [default: NULL]}

\item{gene.labs}{boolean specifying whether gene labels should be plotted [default: TRUE]}

\item{samp.labs}{boolean specifying whether sample labels should be plotted [default: FALSE]}
}
\description{
plot heatmap with gene expression of classifier genes as well as classifications, their probabilities and phenotypic information (mutations)
}
\author{
Robert Piskol
}
