% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getAlpha}
\alias{getAlpha}
\title{get alpha parameter}
\usage{
getAlpha()
}
\value{
alpha parameter that was used for classifier training
}
\description{
get alpha parameter
}
\author{
Robert Piskol
}
