% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getCols}
\alias{getCols}
\title{get colors}
\usage{
getCols(cat = c("COL_MUT", "COL_MUT_BK", "COL_SUB", "HEATMAP_COL3",
  "HEATMAP_BKS3", "CBBPALETTE"))
}
\arguments{
\item{cat}{the category for which to return colors}
}
\value{
vector of colors or color breaks depending on what category was requested
}
\description{
get colors
}
\author{
Robert Piskol
}
