% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getCvfitList}
\alias{getCvfitList}
\title{get list of objects generated using the cv.glmnet function for multiple values of alpha}
\usage{
getCvfitList(idtype = c("Entrez", "Symbol"))
}
\arguments{
\item{idtype}{one of "Entrez" or "Symbol" (default: "Entrez")}
}
\value{
list of cv.glmnet objects at different alpha levels
}
\description{
get list of objects generated using the cv.glmnet function for multiple values of alpha
}
\author{
Robert Piskol
}
