% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getExpressionDataSet}
\alias{getExpressionDataSet}
\title{load expression data}
\usage{
getExpressionDataSet(dataset = showAvailableExpressionDataSets())
}
\arguments{
\item{dataset}{string specifying which data set should be retrieved. Use showAvailableExpressionDataSets() to list available expression data  [default: "CIT"]}
}
\value{
ExpressionSet with the requested data
}
\description{
load expression data
}
\author{
Robert Piskol
}
