% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getFit}
\alias{getFit}
\title{get glmnet object with the trained classifier}
\usage{
getFit(idtype = c("Entrez", "Symbol"))
}
\arguments{
\item{idtype}{one of "Entrez" or "Symbol" (default: Entrez)}
}
\value{
glmnet fit object
}
\description{
get glmnet object with the trained classifier
}
\author{
Robert Piskol
}
