% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getGgColorHue}
\alias{getGgColorHue}
\title{reproduce ggplot's color hues}
\usage{
getGgColorHue(n)
}
\arguments{
\item{n}{the number of colors to generate}
}
\value{
vector of colors
}
\description{
reproduce ggplot's color hues
}
\author{
Robert Piskol
}
