% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getLambda}
\alias{getLambda}
\title{get value of the penalty parameter lambda from crossvalidation for a specific alpha value}
\usage{
getLambda(cvfitList, alpha, s.stat = c("lambda.min", "lambda.1se"))
}
\arguments{
\item{cvfitList}{list of cv.glmnet objects at different alpha levels}

\item{alpha}{value of the alpha parameter for which a specific lambda should be retrieved}

\item{s.stat}{either "lambda.min" or "lambda.1se" (default "lambda.min")}
}
\value{
list of cv.glmnet objects at different alpha levels
}
\description{
get list of objects generated using the cv.glmnet function for multiple values of alpha
}
\author{
Robert Piskol
}
