% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getSignatureGenesPerCMS}
\alias{getSignatureGenesPerCMS}
\title{extract feature (gene) names/ids for a specific CMS subtype from a cvfit file}
\usage{
getSignatureGenesPerCMS(cms, idtype = c("Entrez", "Symbol"),
  s.stat = "lambda.min")
}
\arguments{
\item{cms}{CMS subtype for which features should be reported. This can be a string "CMS1", "CMS2", "CMS3", "CMS4" or numeric value eg. 1,2,3,4 [default: none]}

\item{idtype}{one of "Entrez" or "Symbol" [default: Entrez]}

\item{s.stat}{Value of the penalty parameter lambda for which features will be reported [default: lambda.min]}
}
\value{
vector of feature names
}
\description{
extract feature (gene) names/ids for a specific CMS subtype from a cvfit file
}
\author{
Robert Piskol
}
