% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{getSignatureGenesPerCMSFromFit}
\alias{getSignatureGenesPerCMSFromFit}
\title{extract feature (gene) names/ids for a specific CMS subtype from a cvfit object}
\usage{
getSignatureGenesPerCMSFromFit(cvfit, cms, s.stat = "lambda.min")
}
\arguments{
\item{cvfit}{cvfit object generated using the cv.glmnet function [default: none]}

\item{cms}{CMS subtype for which features should be reported. This can be a string "CMS1", "CMS2", "CMS3", "CMS4" or numeric value eg. 1,2,3,4 [default: none]}

\item{s.stat}{Value of the penalty parameter lambda for which features will be reported [default: lambda.min]}
}
\value{
vector of feature names
}
\description{
extract feature (gene) names/ids for a specific CMS subtype from a cvfit object
}
\author{
Robert Piskol
}
