% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{logIfNeeded}
\alias{logIfNeeded}
\title{apply log2 scale to matrix if needed}
\usage{
logIfNeeded(ex, verbose = FALSE)
}
\arguments{
\item{ex}{a matrix of values that will be log2 transformed (default: none)}

\item{verbose}{specify whether messages should be printed (default: FALSE)}
}
\value{
matrix
}
\description{
apply log2 scale to matrix if needed
}
\author{
Robert Piskol
}
