% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_plotting.R
\name{pieChart}
\alias{pieChart}
\title{plot pie chart for CMS classification}
\usage{
pieChart(df, var = "CMS.glmnet", digits = 1, main = "",
  labTotalN = FALSE)
}
\arguments{
\item{df}{data frame with column that will be used to plot the pie chart [default: none]}

\item{var}{string specifying which column from the supplied data frame to plot [default: "CMS.glmnet"]}

\item{digits}{integer specifying to how many digits pie labels should be rounded [default: 1]}

\item{main}{string specifying the title of the plot [default: ""]}

\item{labTotalN}{boolean specifying whether the total number of samples in each category should be added in addition to the percentage [default: FALSE]}
}
\value{
vector with percentages of samples in each category
}
\description{
plot pie chart for CMS classification
}
\author{
Robert Piskol
}
