% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_plotting.R
\name{plotContingencyGeneral}
\alias{plotContingencyGeneral}
\title{plot pie chart for CMS classification}
\usage{
plotContingencyGeneral(x, y, xlab = NULL, ylab = NULL, main = NULL)
}
\arguments{
\item{x}{vector of values that will be plotted on the x-axis of the contingency table [default: none]}

\item{y}{vector of values that will be plotted on the y-axis of the contingency table. Needs to be sorted in the same way as x [default: none]}

\item{xlab}{x-axis label [default: NULL]}

\item{ylab}{y-axis label [default: NULL]}

\item{main}{string specifying the title of the plot [default: ""]}
}
\value{
vector with percentages of samples in each category
}
\description{
plot pie chart for CMS classification
}
\author{
Robert Piskol
}
