% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{randomString}
\alias{randomString}
\title{generate random string}
\usage{
randomString(length = 10)
}
\arguments{
\item{length}{integer specifying the length of the string to be generated (default: 10)}
}
\value{
string
}
\description{
generate random string
}
\author{
Robert Piskol
}
