% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{showAvailableExpressionDataSets}
\alias{showAvailableExpressionDataSets}
\title{list available expression data}
\usage{
showAvailableExpressionDataSets()
}
\value{
vector listing the expression data sets that are available
}
\description{
list available expression data
}
\author{
Robert Piskol
}
