% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{subsetSamplesByName}
\alias{subsetSamplesByName}
\title{subset samples from the results of the CMS classification using sample names}
\usage{
subsetSamplesByName(x, samples)
}
\arguments{
\item{x}{result of the CMS classification [default: none]}

\item{samples}{vector of sample ids that will be selected from the CMS classification result [default: NULL]}
}
\value{
subsetted result of the CMS classification
}
\description{
subset samples from the results of the CMS classification using sample names
}
\author{
Robert Piskol
}
