% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{subsetSamplesBySignificance}
\alias{subsetSamplesBySignificance}
\title{subset samples from the results of the CMS classification based on a threshold value for the prection}
\usage{
subsetSamplesBySignificance(x, thresh)
}
\arguments{
\item{x}{result of the CMS classification [default: none]}

\item{thresh}{numeric value specifying the threshold that will be used to select samples [default: none]}
}
\value{
subsetted result of the CMS classification
}
\description{
subset samples from the results of the CMS classification based on a threshold value for the prection
}
\author{
Robert Piskol
}
