% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMString_helperFunctions.R
\name{zscoreVals}
\alias{zscoreVals}
\title{z-score expression matrix}
\usage{
zscoreVals(x, center = NA)
}
\arguments{
\item{x}{a matrix of values that will be z-scored (default: none)}

\item{center}{integer value to which to center the matrix to (default: NA - matrix will be centered to mean of all values)}
}
\value{
matrix
}
\description{
z-score expression matrix
}
\author{
Robert Piskol
}
